+++ files/gitlab-cookbooks/gitlab/libraries/gitlab_shell.rb
@@ -29,20 +29,27 @@ module GitlabShell
 
       Gitlab['gitlab_shell']['git_data_directories'] ||=
         if git_data_dirs.any?
-          git_data_dirs
+          Hash[git_data_dirs.map do |name, data_directory|
+            if data_directory.is_a?(String)
+              Chef::Log.warn "Your git_data_dirs settings are deprecated. Please refer to https://docs.gitlab.com/omnibus/settings/configuration.html#storing-git-data-in-an-alternative-directory for updated documentation."
+              [name, { 'path' => data_directory }]
+            else
+              [name, data_directory]
+            end
+          end]
         else
-          { 'default' => git_data_dir }
+          { 'default' => { 'path' => git_data_dir } }
         end
 
       Gitlab['gitlab_rails']['repositories_storages'] ||=
-        Hash[Gitlab['gitlab_shell']['git_data_directories'].map do |name, path|
-          [name, File.join(path, 'repositories')]
+        Hash[Gitlab['gitlab_shell']['git_data_directories'].map do |name, data_directory|
+          [name, { 'path' => File.join(data_directory['path'], 'repositories') }]
         end]
 
       # Important: keep the satellites.path setting until GitLab 9.0 at
       # least. This setting is fed to 'rm -rf' in
       # db/migrate/20151023144219_remove_satellites.rb
-      Gitlab['gitlab_rails']['satellites_path'] ||= File.join(Gitlab['gitlab_shell']['git_data_directories']['default'], "gitlab-satellites")
+      Gitlab['gitlab_rails']['satellites_path'] ||= File.join(Gitlab['gitlab_shell']['git_data_directories']['default']['path'], "gitlab-satellites")
     end
 
     def parse_auth_file